# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/allin/"
        }
    ],
    "paths": {
        "/api/v1/send-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send Otp",
                "description": "Send Otp",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Enter Country Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "+91"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Enter Mobile Number",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "9876543210"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/verify-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify Otp",
                "description": "Verify Otp",
                "operationId": "verifyOtp",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Enter Country Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "+91"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Enter Mobile Number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "9876543210"
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "description": "Enter OTP",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "123456"
                    },
                    {
                        "name": "device_token",
                        "in": "query",
                        "description": "Enter Device Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/check-mobile-exists": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Check Mobile Exists",
                "description": "Check Mobile Exists",
                "operationId": "checkMobileExists",
                "parameters": [
                    {
                        "name": "country_code",
                        "in": "query",
                        "description": "Enter Country Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "+91"
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "Enter Mobile Number",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": "9876543210"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/user-registration": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Registration",
                "description": "User Registration",
                "operationId": "userRegistration",
                "requestBody": {
                    "description": "User Registration Request",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "country_code",
                                    "mobile",
                                    "otp",
                                    "device_token"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "Enter First Name",
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "last_name": {
                                        "description": "Enter Last Name",
                                        "type": "string",
                                        "example": "User"
                                    },
                                    "country_code": {
                                        "description": "Enter Country Code",
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobile": {
                                        "description": "Enter Mobile Number",
                                        "type": "number",
                                        "example": "9876543210"
                                    },
                                    "otp": {
                                        "description": "Enter OTP",
                                        "type": "number",
                                        "example": "123456"
                                    },
                                    "device_token": {
                                        "description": "Enter Device Token",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "profile": {
                                        "description": "Profile Image",
                                        "type": "file"
                                    },
                                    "cover_image": {
                                        "description": "Cover Image",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "device_token",
                        "in": "query",
                        "description": "Enter Device Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/users-mobile-numbers": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Mobile Numbers",
                "description": "User Mobile Numbers",
                "operationId": "userMobileNumbers",
                "responses": {
                    "200": {
                        "description": "json schema",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Invalid Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": ""
        }
    ]
}